requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosmtplib==5.0.0
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
backoff==2.2.1
bcrypt==3.2.2
black==25.11.0
boto3==1.42.8
botocore==1.42.8
ccxt==4.5.24
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coinbase-advanced-py==1.8.2
coincurve==21.0.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
isort==7.0.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycares==4.11.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.3.0
PyYAML==6.0.3
qrcode==8.2
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rsa==4.9.1
s3transfer==0.16.0
sentry-sdk==2.43.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
stripe==14.0.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==13.1
wheel==0.45.1
yarl==1.22.0
